requirements.txt
Flask
Flask-SQLAlchemy
flask_marshmallow

marshmallow
marshmallow-sqlalchemy

psycopg2
python-dotenv